Address review: extract queue file, add Steer/FollowUp to Runtime, return error
- Move QueuedMessage, MessageQueue interface, and inMemoryMessageQueue
  to dedicated message_queue.go file
- Add Steer() and FollowUp() to the Runtime interface — all runtimes
  implement them, no need for a separate MessageInjector interface
- Return error instead of bool from Steer/FollowUp for richer failure
  information (queue full, no active session, network errors)
- Simplify App and SessionManager: call runtime.Steer/FollowUp directly
  without type assertions
